Implementation test_routers refactor
